Add Overcooked game logic header
- Introduced overcooked_logic.h to encapsulate game logic functions for the Overcooked environment, including dish evaluation, grid parsing, and agent interactions.
- Moved relevant function definitions from overcooked.h to the new header to improve code organization and maintainability.
- Updated overcooked.h to include overcooked_logic.h, removing redundant function declarations.
